+++ routes/apiRoutes.js
@@ -2,16 +2,7 @@ var db = require("../models");
 var passport = require("../config/passport");
 
 module.exports = function(app) {
-  // Get most recent (default setting)
-  app.get("/api/entries", function(req, res) {
-    db.Entries.findAll({}).then(function(data) {
-      var hbsObject = {
-        entries: data
-      };
-      console.log(hbsObject);
-      res.render("index", hbsObject);
-    });
-  });
+ 
   // Login User
   // Using the passport.authenticate middleware with our local strategy.
   // If the user has valid login credentials, send them to the members page.
