+++ views/layouts/main.handlebars
@@ -13,6 +13,7 @@
   <link href="https://fonts.googleapis.com/css?family=Alice|Cormorant+Unicase|Vollkorn+SC" rel="stylesheet">
   <!-- icon styles here -->
   <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css" integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU" crossorigin="anonymous">
+  <link rel="stylesheet" href="styles/styles.css">
   <!--jQuery script -->
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
       crossorigin="anonymous"></script>
