+++ views/partials/entries/story-block.handlebars
@@ -5,7 +5,7 @@
       <div class="card-body">
             <h4 class="card-title">{{title}}</h4>
             <p class="card-text">{{entry}}</p>
-            <a href="#" class="btn btn-primary"></a>
-            <span><i class="far fa-heart"></i> </span>
+            <a href="#" class="btn btn-primary">Contribute to the Story</a>
+            <span><i class="far fa-heart"></i></span>
       </div>
